+++ CHANGELOG.md
@@ -1,9 +1,12 @@
+## 0.0.5
+- Add GraphFocus to replace Focus
+
 ## 0.0.4
 
-* Add DraggableFlowGraphView to support draggable layout graph.
-* Abstract GraphView class.
+- Add DraggableFlowGraphView to support draggable layout graph.
+- Abstract GraphView class.
 
 
 ## 0.0.1
 
-* Finish simple DAG layout
+- Finish simple DAG layout